add condition when delete

+++ app/Http/Controllers/Admin/BirthdayEvoucherController.php
@@ -1078,6 +1078,16 @@ class BirthdayEvoucherController extends Controller
             if (!$reward) {
                 return response()->json(['status' => 'error', 'message' => 'Reward not found'], 404);
             }
+
+            $walletExists = UserWalletVoucher::where('reward_id', $reward->id)->exists();
+
+            if ($walletExists) {
+                return response()->json([
+                    'status' => 'error',
+                    'message' => 'This reward exists in user wallet. You cannot delete it.'
+                ], 404);
+            }
+
         
             if ($reward->voucher_image && file_exists(public_path('uploads/image/' . $reward->voucher_image))) {
                 unlink(public_path('uploads/image/' . $reward->voucher_image));
