+++ app/Http/Controllers/Admin/RewardController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Helpers\AdminLogger;
 use App\Http\Controllers\Controller;
 use App\Models\API\GetSRPMerchandiseItemList;
+use App\Models\CartItem;
 use App\Models\Category;
 use App\Models\ClubLocation;
 use App\Models\Merchant;
@@ -1990,6 +1991,22 @@ class RewardController extends Controller
                 return response()->json(['status' => 'error', 'message' => 'Reward not found'], 404);
             }
         
+            $walletExists = UserWalletVoucher::where('reward_id', $reward->id)->exists();
+            
+            if ($walletExists) {
+                return response()->json([
+                    'status' => 'error',
+                    'message' => 'This reward exists in user wallet. You cannot delete it.'
+                    ], 404);
+            }
+                    
+            $cartitem = CartItem::where('voucher_id', $reward->id)->exists();
+            if ($cartitem) {
+                return response()->json([
+                    'status' => 'error',
+                    'message' => 'This reward exists in user cart. You cannot delete it.'
+                    ], 404);
+            }
             if ($reward->voucher_image && file_exists(public_path('uploads/image/' . $reward->voucher_image))) {
                 unlink(public_path('uploads/image/' . $reward->voucher_image));
             }
@@ -2009,6 +2026,7 @@ class RewardController extends Controller
             RewardParticipatingMerchantLocationUpdate::where('reward_id', $reward->id)->delete();
             RewardLocationUpdate::where('reward_id', $reward->id)->delete();        
             $reward->delete();
+            
             AdminLogger::log('delete', Reward::class, $id);
             DB::commit();
